# docs/swagger-openapi.yml
openapi: 3.0.1
info:
  title: Tài liệu REST API Nexora-Commerce
  description: Tài liệu RESTful API hiệu năng cao cho hệ thống thương mại điện tử
    Nexora-Commerce phục vụ tích hợp frontend.
  contact:
    name: Nexora Core Dev Team
    email: dev@nexoracommerce.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Máy chủ phát triển cục bộ (Local Development Server)
security:
- bearerAuth: []
...
paths:
...
  /api/v1/inventories/{productId}/stock-reservations:
    post:
      tags:
      - Quản lý Kho hàng
      summary: Giữ kho cho đơn hàng chờ (Chỉ Admin)
      description: Yêu cầu quyền ADMIN. Tạm giữ một số lượng hàng cụ thể từ kho khả
        dụng cho một đơn hàng đang chờ xử lý. Số lượng bị giữ sẽ không hiển thị bán
        cho người khác.
      operationId: reserveStock
      parameters:
      - name: productId
        in: path
        description: Mã sản phẩm duy nhất
        required: true
        schema:
          type: string
        example: P001
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStockRequest'
        required: true
      responses:
        '403':
          description: Không đủ quyền hạn (yêu cầu vai trò ADMIN)
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        '400':
          description: Không đủ hàng tồn kho hoặc yêu cầu không hợp lệ
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        '404':
          description: Không tìm thấy sản phẩm
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        '401':
          description: Mã xác thực JWT thiếu hoặc không hợp lệ
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        '200':
          description: Giữ kho thành công
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
      security:
      - bearerAuth: []
  /api/v1/inventories/{productId}/stock-releases:
    post:
      tags:
      - Quản lý Kho hàng
      summary: Hoàn trả hàng giữ kho về kho khả dụng (Chỉ Admin)
      description: Yêu cầu quyền ADMIN. Hoàn trả số lượng hàng đang bị tạm giữ về
        lại kho khả dụng. Sử dụng khi đơn hàng bị hủy hoặc giao dịch thất bại.
      operationId: releaseStock
      parameters:
      - name: productId
        in: path
        description: Mã sản phẩm duy nhất
        required: true
        schema:
          type: string
        example: P001
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStockRequest'
        required: true
      responses:
        '403':
          description: Không đủ quyền hạn (yêu cầu vai trò ADMIN)
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        '400':
          description: Yêu cầu giải phóng kho không hợp lệ hoặc không đủ số lượng
            đã giữ
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        '200':
          description: Giải phóng hàng giữ kho thành công
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        '404':
          description: Không tìm thấy sản phẩm
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        '401':
          description: Mã xác thực JWT thiếu hoặc không hợp lệ
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
      security:
      - bearerAuth: []
...
  /api/v1/inventories/{productId}/stock:
    get:
      tags:
      - Quản lý Kho hàng
      summary: Lấy số lượng tồn kho của sản phẩm
      description: Lấy số lượng tồn kho khả dụng thực tế của một sản phẩm (tổng số
        lượng tồn kho trừ đi số lượng đã đặt giữ và đã bán).
      operationId: getStock
      parameters:
      - name: productId
        in: path
        description: Mã sản phẩm duy nhất
        required: true
        schema:
          type: string
        example: P001
      responses:
        '200':
          description: Lấy số lượng tồn kho thành công
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInteger'
        '404':
          description: Không tìm thấy sản phẩm
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInteger'
    patch:
      tags:
      - Quản lý Kho hàng
      summary: Cập nhật số lượng tồn kho (Chỉ Admin)
      description: Yêu cầu quyền ADMIN. Điều chỉnh thủ công số lượng tồn kho của một
        sản phẩm nhằm điều chỉnh kho hoặc nhập thêm hàng.
      operationId: updateStock
      parameters:
      - name: productId
        in: path
        description: Mã sản phẩm duy nhất
        required: true
        schema:
          type: string
        example: P001
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStockRequest'
        required: true
      responses:
        '403':
          description: Không đủ quyền hạn (yêu cầu vai trò ADMIN)
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        '400':
          description: Yêu cầu cập nhật không hợp lệ hoặc số lượng không chính xác
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        '404':
          description: Không tìm thấy sản phẩm
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        '401':
          description: Mã xác thực JWT thiếu hoặc không hợp lệ
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        '200':
          description: Cập nhật tồn kho thành công
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
      security:
      - bearerAuth: []
...
  /api/v1/inventories/{productId}:
    get:
      tags:
      - Quản lý Kho hàng
      summary: Lấy chi tiết thông tin kho của sản phẩm
      description: Truy xuất chi tiết dữ liệu kho của một sản phẩm bao gồm số lượng
        tồn thực tế, số lượng đã đặt giữ, số lượng đã bán và cấu hình kho.
      operationId: getInventoryDetails
      parameters:
      - name: productId
        in: path
        description: Mã sản phẩm duy nhất
        required: true
        schema:
          type: string
        example: P001
      responses:
        '200':
          description: Lấy chi tiết tồn kho thành công
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInventoryResponse'
        '404':
          description: Không tìm thấy sản phẩm
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInventoryResponse'
  /api/v1/inventories/{productId}/availability:
    get:
      tags:
      - Quản lý Kho hàng
      summary: Kiểm tra tồn kho khả dụng cho một số lượng cụ thể
      description: Xác minh xem tồn kho hiện tại của sản phẩm có đáp ứng được số lượng
        mua mong muốn hay không.
      operationId: hasEnoughStock
      parameters:
      - name: productId
        in: path
        description: Mã sản phẩm duy nhất
        required: true
        schema:
          type: string
        example: P001
      - name: quantity
        in: query
        description: Số lượng cần kiểm tra tính khả dụng
        required: true
        schema:
          type: integer
          format: int32
        example: 5
      responses:
        '400':
          description: Tham số số lượng không hợp lệ
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseBoolean'
        '200':
          description: Kiểm tra trạng thái khả dụng thành công
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseBoolean'
        '404':
          description: Không tìm thấy sản phẩm
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseBoolean'
...
components:
  schemas:
...
    PaginationInfo:
      type: object
      properties:
        page:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
...
    ApiResponseVoid:
      type: object
      properties:
        success:
          type: boolean
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        message:
          type: string
        data:
          type: object
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
...
    UpdateStockRequest:
      required:
      - quantity
      type: object
      properties:
        quantity:
          minimum: 1
          type: integer
          format: int32
...
    ApiResponseInteger:
      type: object
      properties:
        success:
          type: boolean
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        message:
          type: string
        data:
          type: integer
          format: int32
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
    ApiResponseBoolean:
      type: object
      properties:
        success:
          type: boolean
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        message:
          type: string
        data:
          type: boolean
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
...
    ApiResponseInventoryResponse:
      type: object
      properties:
        success:
          type: boolean
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: '#/components/schemas/InventoryResponse'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
    InventoryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        productId:
          type: string
        quantity:
          type: integer
          format: int32
        reservedQuantity:
          type: integer
          format: int32
        soldQuantity:
          type: integer
          format: int32
        availableQuantity:
          type: integer
          format: int32
...
  securitySchemes:
    bearerAuth:
      type: http
      description: Please paste your JWT Access Token in the box below to authorize.
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT